+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This file is used to list changes made in each version of the ruby_build cookboo
 - Fix YAML
 - Use platform? helper in the attributes file
 - Remove the unnecessary long_description field in metadata.rb
+- Fix libgdbm package name in attributes for debian 10
+- Add debian-10 platform to test kitchen configurations
 
 ## 1.2.0 (2019-23-01)